Revert "[operations #1] refactor operation to package, add fee param"
