+++ library/src/main/java/ch/decent/sdk/api/MessagingApi.kt
@@ -6,13 +6,12 @@ import ch.decent.sdk.DCoreApi
 import ch.decent.sdk.crypto.Credentials
 import ch.decent.sdk.model.Account
 import ch.decent.sdk.model.ChainObject
-import ch.decent.sdk.model.Fee
 import ch.decent.sdk.model.Memo
 import ch.decent.sdk.model.Message
 import ch.decent.sdk.model.MessagePayload
 import ch.decent.sdk.model.MessagePayloadReceiver
 import ch.decent.sdk.model.MessageResponse
-import ch.decent.sdk.model.operation.SendMessageOperation
+import ch.decent.sdk.model.SendMessageOperation
 import ch.decent.sdk.model.TransactionConfirmation
 import ch.decent.sdk.net.model.request.GetMessageObjects
 import io.reactivex.Single
@@ -98,11 +97,9 @@ class MessagingApi internal constructor(api: DCoreApi) : BaseApi(api) {
    *
    * @return send message operation
    */
-  @JvmOverloads
   fun createMessageOperation(
       credentials: Credentials,
-      messages: List<Pair<ChainObject, String>>,
-      fee: Fee = Fee()
+      messages: List<Pair<ChainObject, String>>
   ): Single<SendMessageOperation> = Single.zip(
       api.accountApi.get(credentials.account),
       Single.merge(messages.map { api.accountApi.get(it.first) }).toList(),
@@ -113,7 +110,7 @@ class MessagingApi internal constructor(api: DCoreApi) : BaseApi(api) {
       MessagePayloadReceiver(recipient.id, msg.message, recipient.options.memoKey, msg.nonce)
     }
     val data = MessagePayload(sender.id, payloadReceivers, sender.options.memoKey)
-    SendMessageOperation(api.core.gson.toJson(data), credentials.account, fee = fee)
+    SendMessageOperation(api.core.gson.toJson(data), credentials.account)
   }
 
   /**
@@ -126,11 +123,8 @@ class MessagingApi internal constructor(api: DCoreApi) : BaseApi(api) {
    */
   fun createMessageOperationUnencrypted(
       credentials: Credentials,
-      messages: List<Pair<ChainObject, String>>,
-      fee: Fee = Fee()
-  ): Single<SendMessageOperation> = Single.just(
-      SendMessageOperation(api.core.gson.toJson(MessagePayload(credentials.account, messages)), credentials.account, fee = fee)
-  )
+      messages: List<Pair<ChainObject, String>>
+  ): Single<SendMessageOperation> = Single.just(SendMessageOperation(api.core.gson.toJson(MessagePayload(credentials.account, messages)), credentials.account))
 
   /**
    * Send messages to receivers
@@ -142,9 +136,8 @@ class MessagingApi internal constructor(api: DCoreApi) : BaseApi(api) {
    */
   fun send(
       credentials: Credentials,
-      messages: List<Pair<ChainObject, String>>,
-      fee: Fee = Fee()
-  ): Single<TransactionConfirmation> = createMessageOperation(credentials, messages, fee).flatMap {
+      messages: List<Pair<ChainObject, String>>
+  ): Single<TransactionConfirmation> = createMessageOperation(credentials, messages).flatMap {
     api.broadcastApi.broadcastWithCallback(credentials.keyPair, it)
   }
 
@@ -158,9 +151,8 @@ class MessagingApi internal constructor(api: DCoreApi) : BaseApi(api) {
    */
   fun sendUnencrypted(
       credentials: Credentials,
-      messages: List<Pair<ChainObject, String>>,
-      fee: Fee = Fee()
-  ): Single<TransactionConfirmation> = createMessageOperationUnencrypted(credentials, messages, fee).flatMap {
+      messages: List<Pair<ChainObject, String>>
+  ): Single<TransactionConfirmation> = createMessageOperationUnencrypted(credentials, messages).flatMap {
     api.broadcastApi.broadcastWithCallback(credentials.keyPair, it)
   }
 }
