+++ library/src/main/java/ch/decent/sdk/model/Transaction.kt
@@ -1,7 +1,6 @@
 package ch.decent.sdk.model
 
 import ch.decent.sdk.crypto.ECKeyPair
-import ch.decent.sdk.model.operation.BaseOperation
 import ch.decent.sdk.model.types.UInt16
 import ch.decent.sdk.model.types.UInt32
 import ch.decent.sdk.net.serialization.Serializer
