+++ library/src/main/java/ch/decent/sdk/model/TypeAdapters.kt
@@ -4,8 +4,6 @@ import ch.decent.sdk.crypto.Address
 import ch.decent.sdk.crypto.Wallet
 import ch.decent.sdk.crypto.address
 import ch.decent.sdk.crypto.dpk
-import ch.decent.sdk.model.operation.BaseOperation
-import ch.decent.sdk.model.operation.EmptyOperation
 import com.google.gson.Gson
 import com.google.gson.TypeAdapter
 import com.google.gson.TypeAdapterFactory
