--- library/src/main/java/ch/decent/sdk/model/operation/EmptyOperation.kt
@@ -1,9 +1,0 @@
-package ch.decent.sdk.model.operation
-
-import ch.decent.sdk.model.Fee
-import ch.decent.sdk.model.OperationType
-
-class EmptyOperation(type: OperationType) : BaseOperation(type, Fee()) {
-
-  override fun toString(): String = type.toString()
-}
