+++ library/src/main/java/ch/decent/sdk/net/model/request/GetRequiredFees.kt
@@ -2,7 +2,7 @@ package ch.decent.sdk.net.model.request
 
 import ch.decent.sdk.DCoreConstants
 import ch.decent.sdk.model.AssetAmount
-import ch.decent.sdk.model.operation.BaseOperation
+import ch.decent.sdk.model.BaseOperation
 import ch.decent.sdk.model.ChainObject
 import ch.decent.sdk.model.ObjectType
 import ch.decent.sdk.net.model.ApiGroup
