+++ library/src/main/java/ch/decent/sdk/net/serialization/Serializer.kt
@@ -20,27 +20,26 @@
 package ch.decent.sdk.net.serialization
 
 import ch.decent.sdk.crypto.Address
+import ch.decent.sdk.model.AccountCreateOperation
 import ch.decent.sdk.model.AccountOptions
+import ch.decent.sdk.model.AccountUpdateOperation
+import ch.decent.sdk.model.AddOrUpdateContentOperation
 import ch.decent.sdk.model.AssetAmount
 import ch.decent.sdk.model.AuthMap
 import ch.decent.sdk.model.Authority
 import ch.decent.sdk.model.ChainObject
 import ch.decent.sdk.model.CustodyData
-import ch.decent.sdk.model.Fee
+import ch.decent.sdk.model.CustomOperation
 import ch.decent.sdk.model.KeyPart
 import ch.decent.sdk.model.Memo
 import ch.decent.sdk.model.PubKey
 import ch.decent.sdk.model.Publishing
+import ch.decent.sdk.model.PurchaseContentOperation
 import ch.decent.sdk.model.RegionalPrice
+import ch.decent.sdk.model.SendMessageOperation
 import ch.decent.sdk.model.Transaction
+import ch.decent.sdk.model.TransferOperation
 import ch.decent.sdk.model.VoteId
-import ch.decent.sdk.model.operation.AccountCreateOperation
-import ch.decent.sdk.model.operation.AccountUpdateOperation
-import ch.decent.sdk.model.operation.AddOrUpdateContentOperation
-import ch.decent.sdk.model.operation.CustomOperation
-import ch.decent.sdk.model.operation.PurchaseContentOperation
-import ch.decent.sdk.model.operation.SendMessageOperation
-import ch.decent.sdk.model.operation.TransferOperation
 import ch.decent.sdk.utils.SIZE_OF_POINT_ON_CURVE_COMPRESSED
 import ch.decent.sdk.utils.unhex
 import okio.Buffer
