+++ library/src/test/java/ch/decent/sdk/api/HistoryApiTest.kt
@@ -28,9 +28,7 @@ class HistoryApiTest(channel: Channel) : BaseApiTest(channel) {
   }
 
   @Test fun `should list relative account history`() {
-    val test = api.historyApi.listOperationsRelative(Helpers.account, limit = 100)
-        .map { it.joinToString("\n") { it.operation.toString() } }
-        .doOnSuccess { System.out.println(it) }
+    val test = api.historyApi.listOperationsRelative(Helpers.account, limit = 10)
         .subscribeOn(Schedulers.newThread())
         .test()
